requirements.txt
fastapi
uvicorn
groq
python-dotenv
httpx
tavily-python